pyproject.toml
[tool.poetry]
name = "spotify-ubercorn"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
typer = "^0.1.1"
python-dotenv = "^0.13.0"
pyfy = "^2.0.7"
starlette = "^0.13.3"
uvicorn = "^0.11.3"
unicornhathd = {version = "^0.0.4", optional = true}
pillow = "^7.1.1"
numpy = {version = "^1.18.3", optional = true}
strawberry-graphql = "^0.23.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
mypy = "^0.770"
black = {version = "^19.10b0", allow-prereleases = true}
pdbpp = "^0.10.2"

[tool.poetry.extras]
unicornhathd = ["unicornhathd", "numpy"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

